+++ tests/core/test_tab.py
@@ -1,0 +1,28 @@
+import zendriver as zd
+
+
+async def test_set_user_agent_sets_navigator_values(browser: zd.Browser):
+    tab = browser.main_tab
+
+    await tab.set_user_agent(
+        "Test user agent", accept_language="testLang", platform="TestPlatform"
+    )
+
+    navigator_user_agent = await tab.evaluate("navigator.userAgent")
+    navigator_language = await tab.evaluate("navigator.language")
+    navigator_platform = await tab.evaluate("navigator.platform")
+    assert navigator_user_agent == "Test user agent"
+    assert navigator_language == "testLang"
+    assert navigator_platform == "TestPlatform"
+
+
+async def test_set_user_agent_defaults_existing_user_agent(browser: zd.Browser):
+    tab = browser.main_tab
+    existing_user_agent = await tab.evaluate("navigator.userAgent")
+
+    await tab.set_user_agent(accept_language="testLang")
+
+    navigator_user_agent = await tab.evaluate("navigator.userAgent")
+    navigator_language = await tab.evaluate("navigator.language")
+    assert navigator_user_agent == existing_user_agent
+    assert navigator_language == "testLang"